requirements.txt
colorlog==6.7.0
homeassistant==2023.9.0
pyserial==3.5.0